Student_Maagement_System/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.8
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
filelock==3.16.1
fonttools==4.54.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
numpy==2.1.1
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pymongo==4.10.1
pyparsing==3.1.4
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.6
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
virtualenv==20.26.6
watchfiles==0.24.0
websockets==13.1
